refactor: put layout components into app
