--- packages/lib/src/components/Layouts/PublicLayout.tsx
@@ -1,9 +1,0 @@
-import { Center } from '@mantine/core'
-
-type Props = {
-  children: React.ReactNode
-}
-
-export function PublicLayout({ children }: Props): JSX.Element | null {
-  return <Center>{children}</Center>
-}
