--- packages/lib/src/components/Layouts/index.ts
@@ -1,2 +1,0 @@
-export * from './AuthLayout'
-export * from './PublicLayout'
