+++ packages/lib/src/components/index.ts
@@ -3,7 +3,6 @@ export * from './DeleteDialog'
 export * from './Footer'
 export * from './Header'
 export * from './Home'
-export * from './Layouts'
 export * from './LoadingSpinner'
 export * from './Login'
 export * from './NavBar'
